+++ javascript/setHints.js
@@ -107,11 +107,12 @@ async function setHints() {
   const t1 = performance.now();
   log('setHints', { type: localeData.type, elements: elements.length, localized, hints, data: localeData.data.length, time: t1 - t0 });
   // sortUIElements();
-  removeSplash();
   // validateHints(elements, localeData.data);
 }
 
+/*
 onAfterUiUpdate(async () => {
   if (localeData.timeout) clearTimeout(localeData.timeout);
   localeData.timeout = setTimeout(setHints, 250);
 });
+*/